feat: add accessibility (#7)
## This PR develops the portfolio project's accessibility so people can
understand, navigate, and interact with the website and people with
disabilities can use them ensuring there are no barriers that prevent
interaction with it .

### List of points checked include :

- **Page title** _`("no issues found")`_ the page title provides a
concise description of the page's purpose, Good page titles are
particularly important for orientation, to help people know where they
are and move between pages open in their browser. The first thing screen
readers say when the user goes to a different web page is the page
title.

- **Image text alternatives (img alt="alt text")** _`('issues fixed')`_
- Alternative (Alt) Text is meant to convey the “why” of the image as it
relates to the content of a document or webpage. It is read aloud to
users by screen reader software, and it is indexed by search engines. It
also displays on the page if the image fails to load, as in this example
of a missing image.

- **Text headings** _`('No issues found')`_ Organizing web pages by
headings helps users get a sense of the page's organization and
structure. Visually, headings are presented as larger and more distinct
than surrounding text. Making texts larger helps guide the eye around
the page.The reader should be able to preview what your page covers,
your argument by reading just the headings.

- **Color contrast** _`('No issues found')`_ Color contrast impacts the
readability of your content on the web and in print. It is especially
important for users who are low vision or for users who are colorblind.
Good color contrast means all users can see your content no matter the
device they're using or the lighting of their surroundings. Although
color may be a great tool to visually convey information, it may also
create accessibility barriers.When using color, it is important to have
sufficient contrast between two colors: `the background color` and
`foreground color`. Low contrast colors will make it difficult for
visual users to read and/or understand the content `(text or images of
text)` on the page.

- **Resize** \_`('No issues found')`\_Many users with limited vision
resize the text, or zoom into pages to make the content easier to read.
This allows them to experience the content on the page without using
screen magnification or screen reader software.Enlarging text is
primarily the browser’s responsibility. Browsers accomplish this chiefly
in one of two ways: enlarging the font size only, or zooming into the
entire page and applying responsive styles, as appropriate.

- **Interaction** _`('No issues found')`_ Many people cannot use a mouse
and rely on the keyboard to interact with the Web. People who are blind
and some sighted people with mobility impairments rely on the keyboard
or on assistive technologies and strategies that rely on keyboard
commands, such as voice input. Accessible websites enable people to
access all content and functionality

- **Forms, labels, and errors** _`('No issues found')`_ When labels are
marked up correctly, people can interact with them using only the
keyboard, using voice input, and using screen readers. Also, the label
itself becomes clickable, increasing the target area and making it
easier to select small radio buttons or checkboxes.

- **Moving content** _`('No issues found')`_ - There is no moving,
flashing, or blinking Content in this page.

- **Multimedia** _`('No issues found')`_ - There is no video or audio in
this webpage.

- **The basic structure of the page** _`('No issues found')`_ There is
no issue with the basic structure of the page

### Here is the link to the projects [requirement
page](https://github.com/bolah2009/curriculum-html-css/blob/main/portfolio/6_accessability.md)